lib, test, tools: compare and update readme table
This adds the ability create a markdown table (using ghmd-table) during runtime
based on the types we create in lib/bootstrap.js and compare that to the current
README.md table.  Also, we may manually update the table with
`node tools/table.js --update`.

This also adds a test file that compares the runtime-created ghmd table to the
current README.md table, if they are not alike (most likely b/c there's been an
new type added in lib/bootstrap.js) the test fails with code `111`.

Note that the ghmd table is only created if we are in 'ch-arge:bootstrap.js'
debug mode on `process.env.DEBUG`.  So it is necessary that the value is set
before we initially require lib/bootstrap.js' or else the test will fail.  Our
solution is to add an extra line in spec/support/jasmine.json under the field
'headers' that runs tools/table.js with no arguments and does nothing more than
set 'ch-arge:boostrap.js' to `process.env.DEBUG`.

Most of this happens in tools; tools/map-table.js uses the `ghmd-table` api to
simplify the creation of the table, tools/table.js is where the compare and
update happens.  `node tools/table.js --compare` simply compares and exits with
status code `111` if there is an update.  `node tools/table.js --update` updates
only if there is an update.  `node tools/table.js` is a NOP aside from setting
adding 'ch-arge:bootstrap.js' to `process.env.DEBUG`.

This is part of #18 suite, getting it ready for #17.

Refs: #18